+++ HISTORY.md
@@ -5,3 +5,7 @@
 
 * v1.0.1
     * setSafearea method: change safearea after object initialization
+
+* v1.0.2
+    * changed the resize logic to not listen on the window resize event, instead poll for width/height changes of the element itself
+    * new option "resizeInterval"